Get SDP working with Chromium RTCPeerConnection
This uses Chromium and the gstreamer-send fiddle from
https://github.com/pion/example-webrtc-applications
to test that the generated SDP is acceptable and that the browser is
happy to attempt a connection to the gathered candidates.

It needs a lot of cleaning up, mainly around the networking, due to the
threading and the way the JoinHandles are passed around, but it's a
start!